infrastructure/deployment-requirements.txt
# Azure Management SDKs
azure-identity==1.13.0
azure-mgmt-resource==23.0.1
azure-mgmt-web==7.0.0
azure-mgmt-cosmosdb==9.2.0
azure-mgmt-signalr==1.1.0
azure-mgmt-applicationinsights==4.0.0
azure-cosmos==4.3.1

# Supporting packages
python-dotenv==1.0.0
cryptography==41.0.1
PyYAML==6.0.1
aiohttp==3.8.4